requirements.txt
pygame-ce
esper
pygbag
pyinstaller
